New Website!
Re-written whole website from the base. Previous Website had 330~ Lines, and looked Ugly, but this one, is New, and has around 700~ Lines. It is totally Black and White MInimalist Portfolio, with all my Projects, and Satisfying and Smooth Transitions, Easter Eggs, Typing Effects, Floating Effects, and a lot more. I don't know, what I have actually added. Made with the help of a friend of mine who is really good at CSS, and JavaScript. I suck anyway :( So yeah, go ahead and check this, I might not update until and unless I am making a new project, or when it's a new year, because I need to change the CopyRight year too! This took us 10 Hours of Thinking and Coding, Debugging, etc. But who cares, enjoy :)